// src/api/errors/BadRequestError.ts
// This file was auto-generated by Fern from our API Definition.

import type * as core from "../../core/index.js";
import * as errors from "../../errors/index.js";

export class BadRequestError extends errors.DeepgramError {
    constructor(body?: unknown, rawResponse?: core.RawResponse) {
        super({
            message: "BadRequestError",
            statusCode: 400,
            body: body,
            rawResponse: rawResponse,
        });
        Object.setPrototypeOf(this, new.target.prototype);
        if (Error.captureStackTrace) {
            Error.captureStackTrace(this, this.constructor);
        }

        this.name = "BadRequestError";
    }
}